pipeline: Convert sink element to omx_videosink (requires gstomx update).
The VPSS video compositor doesn't appear to be supported via the V4L2 sink,
so we need to convert the sink elements to make use of the omx_videosink
element. However, there were a few bugs in the LCD setup hardcoded in gstomx
that require an update.

See 435afc40161652ebbfacdad019c7c6f7080b49f1 in the chronos-components repo.