Gate delivery of SIGALRM to allow clock stretching of playback timer.
When the playback timer approaches the maximum framerate of the FPGA
we observe jittery playback, this appears to be caused by playback
timer and FPGA framerate falling in and out of phase with each other.
To fix this, we stretch the playback timer until the FPGA framesync
signal to keep the two in phase.