dng: Add NEON accelerated memcpy, as part of a fix for cache corruption.
Enabling the data cache in the Cortex-A8 framebuffer region has ended in tears,
and leads to video corruption in the H.264 encoder. As a workaround, we go back
to disabling the cache, and instead use a little help from NEON to efficiently
get the video data out of the framebuffer region and into cacheable memory.