Ignore SIGHUP while in recording mode, and work around h264 framerate bug.
HDMI hotplug events generate a SIGHUP, which can cause a recording event
to abort. So we choose to ignore them while recording in order to prevent a
loss of data.

We also address a bug where the mp4mux element sets a bogus framerate value
if the first frame of a recording is dropped. Unfortuately due to another bug
in OMX, the first frame will always junk data. So we're caught between two
issues.