Minor tweak to pipeline flow control to improve throughput.
In order to prevent dropped frames on slow media, we need to check
the OMX buffer level before drawing the next frame, unfortunately this
incurs an extra round-trip to the co-processor for every frame. When the
frame level is known to be high, we can pipeline this check to speed up
the video playback a bit.