Fix pipeline crash from unsafe pthread signal handling in OMX.
The use of signals to seek through playback would cause OMX to crash
during signal handling. To ensure safe playback with OMX, use
pthread_sigmask() to ensure that the playback signals are only ever
received by the playback_thread.

The handling of SIGHUP by the main thread could also lead to an OMX
crash. So there is still more cleanup to be done to ensure thread
safety concerns are met.