Fix pixel bit-packing in RAW12 format.
RAW12 format has been broken for some time due to a combination of bugs in
the OMX camera's processing of raw RGB888 pixel data. The first problem was
that the camera element would swap the R and B channels causing a byte
reflection. The second, and more serious problem is due to a reclocking
issue in the FPGA, the video would inconsistently drop the first pixel.
This was most noticeable as colour corruption after debayering.

We opt to fix the issue by configuring the FPGA for RAW16 mode, and use
some NEON instructions to repack the pixels back into RAW12 format when
writing to disk.