Fix race condition in generation of DBus segment signal.
It turns out that D-Bus is not thread-safe, and the generation of a
D-Bus signal from outside of the Glib main loop occasionally leads
to crashes and data corruption in D-Bus. To resolve this, pass the
generation of the segment signal to the main loop by registering an
one-shot idle source.