pipeline: Remove playback timer and minimize display reconfig events.
It seems that the some video deadlocks are being caused by a race
condition between the playback seeking logic and the switching of
display config. We attempt to fix this by moving the seeking logic
and the display config together into the playback thread.

This requires the addition of a pipe to pass commands into the
playback thread when a change is required.