Fix DNG/TIFF save after adding the NEON flip.
The NEON flip element only really works for NV12 video data, which
isn't what we get when in raw mode, leading to a pipeline crash.
Lucky for us, this can be fixed by some minor shuffling of the DNG
and TIFF tags to change the video orientation in the viewer.