Added handling of start and stop commands.
Saves live recordings to the file specified in the JSON format,
no longer starts recording automatically when rest of the pipeline
is started and will require a command from cam-json or a d-bus
call. Audio has been disabled as it causes video to freeze up
after a few minutes. It will be reimplemented when fixed.

See the updated README.md for more details on liverecord.