Fix g_value_init assertion in D-Bus type conversion.
g_value_init was throwing an assert on Debian builds when converting
GValues from one type to another. This was due to uninitialized stack
junk causing g_value_init to mistake the GValue as being already
initialized as some other type. Clearing the structure with memset
first fixes the issue.

Note that we can't use the G_VALUE_INIT macro, since that isn't present
on the Arago builds.