Add videoSegments property to the video API.
By reading the videoSegments property, the UI and control API can get
a detailed map of the recorded video and how it is broken down by
recording trigger events. For FPGA versions 3.22 and newer we also add
signal generation on segment update. Earlier versions of the FPGA had
bugs that could generate an overwhelming quantity of segment data.