Add RTSP session hook for dynamic client connection.
The RTSP server code can use the session hook to notify the rest of the
daemon about the negotiation state of the RTSP sessions. The h.264
encoder can then use this hook to update the client list for the RTP/UDP
sink.

This is still limited to a single session for the time being.