Improved error handling
Updated error messages to be appropriate for the UI,
changed live record framerate back to 60fps. Added
error handling for fdsink element errors, but it does
not handle errors and restart the pipeline reliably
at the moment.

TODO: Reliably handle fdsink errors, ie. when a user
disconnects media in the middle of a recording.