pcUtil: Use condition variables for data signaling.
When pcUtil needs to wait for data, it used to do so in a busy loop
that could consume the entire CPU if there is a large backlog of
messages to process. Using a condition variable to signal the arrival
of data lets the CPU switch to other tasks until the PMIC firmware
responds.

Closes: https://github.com/krontech/chronos-cli/issues/6