OMX deadlock workaround due to VIP parser overflow.
The VIP parser in the dm8148 can deadlock due to DDR memory starvation when
a parser overflow occurs, such overflows are known to occur while starting
and stopping the omx_camera element and active video is being played out
of the FPGA. The dm814x eratta document claims that the EZSDK includes a
software workaround for this, but closer examination of the workaround code
reveals a race condition where this can still occur during setup.

To try and do a better job of avoiding these deadlocks, ensure that the
video stream from the FPGA is halted during pipeline setup by delaying the
video start until after the omx_camera element reaches the PLAYING state.

See Advisory 3.0.16 from the DM814x eratta document (SPRZ343C) for more
information, disappointment and sadness.