pcUtil: Fix polling bugs leading to choppy performance.
Incorrect use of select() and non-blocking IO lead to some performance
deficiencies in pcUtil. The first issue was that the select call for the
master socket unnecessarily checking all client sockets, leading to high
CPU usage when any client socket was ready. The second issue was that
the client thread would occasionally block, leading to inconsistent
latency.