pcUtil: Bug in powerdown would lead to hard shutdown.
When pcUtil would detect the shutdown request flag from the PMIC, it
would respond to the PMIC with a shutdown command, causing the camera
to immediately power down the voltage regulators without giving the
camera a change to synch any changes to storage.

The handling of the shutdown command is already implemented in the
kernel, so we really don't need to do anything to correctly shutdown.