Add parameter for the overlay position.
The overlay enable was being saved persistently as a saved parameter
in the video system, but the position of the textbox was not. This
tended to result in the overlay magically appearing at the wrong
location on the screen. By adding the overlayPosition as a saved
parameter we should at least keep it somewhat consistent.