Add timing block and CLI filter options to cam-regdump.
After adding the register parsing for the FPGA timing block, the regdump
tool is starting to get extremely verbose, so we also some getopt code
to allow the user to filter out just a single register block. When no
options are provided, the default behavior is to print all register
blocks.