Re-implement zebra parameter in video API calls.
Zebra stripe configuration was partially moved into a D-Bus parameter
in commit 37ed8fa93740b6940d2579bf12e523157c3d1624, but this change
broke the ability of camApp to control zebras through the configure()
and livedisplay() calls.

Some handholding is needed when dealing with this parameter because it
was switched from a boolean to a float in the parameter change, but we
don't want to overwrite the magnitude unecessarily.

See: https://github.com/krontech/chronos-cam-app/issues/143