Add SCGI daemon for D-Bus to web service conversion.
This is yet-another web API binding, but this one is intended to
operate as the back-end behind an HTTP server using the Simple-CGI
protocol. This tools accepts HTTP POSTs in application/json format
to perform API calls, and also allows a GET on /subscribe to get
an HTML5 SSE stream with the camera's asynchronos events.