scgi: Add support for large payloads.
Receiveing large payloads is performed by optionally allocating a buffer
for the request body at the end of the REQ_HEADERS state if the contentlen
exceeds the space available in the inline buffer. The maximum request size
in this case is limited to 1MB.

Transmitting large payloads is performed by adding a new scgi_take_payload()
method that takes ownership of a heap-allocated buffer. The buffer will be
automatically freed when closing the SCGI session.