scgi: Also accept query string parameters for API calls.
This turns out to be really handy for debugging since you can just type the
parameters into a browser and get a result. This only supports primitive
JSON types, so you can't invoke any API that requires nested JSON in it (you
should be using a POST for that case anyways).

We also start work on a 'describe' URL that generates introspection data for
the API documentation. So far this just returns the contents of the
availableKeys call.