Add crop/padding to source video frame.
This replaces the neonflip element with a neoncrop element to apply
frame padding and cropping operations at the start of the video
pipeline. This enables the video system to work with source video
data that might now otherwise fit with the downstream elements (eg:
h.264 compression and the OMX scalers).

As a side-effect, this means that we no longer support flipping
video (which never quite worked in the first place), and we have
to track separately the size video in the stream from the video in
the FPGA.