neoncrop: Add explicit cache flush after transform.
Since the video system is typically full of hardware accelerators, we
must ensure that changes to the video frames are properly flushed to
memory, otherwise frame corruption may occur.

Unfortunately, the system call to perform such a cache flush is not
universally exported by all architectures, so this entails manually
invoking the call in assembly.