Add support for automounting FUSE filesystems
Filesystems which use FUSE can't be mounted directly from a udev
script, since they require a backend process to keep running after
the udev script exits. The recommended way to deal with this is to
launch a systemd service to manage the mounts.

This should enable a camera to support exotic filesystems by simply
installing their corresponding FUSE filesystem package.