pcUtil: Updated power manager firmware to v10.
A delay has been added to account for users who might hold
the power button down for longer than 500 ms when turning their
cameras on. The camera will allow for the power button to be held
for up to 3 seconds when powering on until interpreting the hold
as a request to turn the camera off.